+++ ricaun.Revit.DB.Shape/ShapeCreator.cs
@@ -21,11 +21,12 @@ namespace ricaun.Revit.DB.Shape
         /// </summary>
         /// <param name="center">The center point of the box.</param>
         /// <param name="scaleRadius">The scale size of the box.</param>
+        /// <param name="graphicsStyleId">The ID of the graphics style to use for the lines. If not specified, the lines will use the default graphics style.</param>
         /// <returns>An array of lines representing the edges of the 3D box.</returns>
-        public static Line[] CreateBoxLines(XYZ center, double scaleRadius)
+        public static Line[] CreateBoxLines(XYZ center, double scaleRadius, ElementId graphicsStyleId = null)
         {
             var scaleXYZ = new XYZ(scaleRadius, scaleRadius, scaleRadius);
-            return CreateBoxLines(center - scaleXYZ, center + scaleXYZ);
+            return CreateBoxLines(center - scaleXYZ, center + scaleXYZ, graphicsStyleId);
         }
 
         /// <summary>
@@ -33,9 +34,10 @@ namespace ricaun.Revit.DB.Shape
         /// </summary>
         /// <param name="min">The minimum point of the box. This point defines the lower corner of the box.</param>
         /// <param name="max">The maximum point of the box. This point defines the upper corner of the box.</param>
+        /// <param name="graphicsStyleId">The ID of the graphics style to use for the lines. If not specified, the lines will use the default graphics style.</param>
         /// <returns>An array of lines representing the edges of the 3D box.</returns>
         /// <remarks>Ignore Lines with distance too short.</remarks>
-        public static Line[] CreateBoxLines(XYZ min, XYZ max)
+        public static Line[] CreateBoxLines(XYZ min, XYZ max, ElementId graphicsStyleId = null)
         {
             // Create an array to store the lines
             Line[] lines = new Line[12];
@@ -55,18 +57,23 @@ namespace ricaun.Revit.DB.Shape
             // Create the lines for each edge of the cube using a for loop
             for (int i = 0; i < 4; i++)
             {
-                lines[i] = CreateBound(vertices[i], vertices[(i + 1) % 4]);
-                lines[i + 4] = CreateBound(vertices[i], vertices[i + 4]);
-                lines[i + 8] = CreateBound(vertices[i + 4], vertices[(i + 1) % 4 + 4]);
+                lines[i] = CreateBound(vertices[i], vertices[(i + 1) % 4], graphicsStyleId);
+                lines[i + 4] = CreateBound(vertices[i], vertices[i + 4], graphicsStyleId);
+                lines[i + 8] = CreateBound(vertices[i + 4], vertices[(i + 1) % 4 + 4], graphicsStyleId);
             }
 
             return lines.OfType<Line>().ToArray();
         }
-        private static Line CreateBound(XYZ point1, XYZ point2)
+        private static Line CreateBound(XYZ point1, XYZ point2, ElementId graphicsStyleId = null)
         {
             try
             {
-                return Line.CreateBound(point1, point2);
+                var line = Line.CreateBound(point1, point2);
+
+                if (graphicsStyleId is not null)
+                    line.SetGraphicsStyleId(graphicsStyleId);
+
+                return line;
             }
             catch { }
             return null;